add tests

+++ src/main/kotlin/com/snakesandladders/domain/Token.kt
@@ -1,3 +1,3 @@
 package com.snakesandladders.domain
 
-data class Token(var id: Int, var color: String, var position: Int)
+data class Token(var id: Long, var color: String, var position: Int)
